% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cjs.sim.r
\name{cjs.sim}
\alias{cjs.sim}
\title{operating model for CJS mark recapture data}
\usage{
cjs.sim(
  phi,
  p,
  nInd = 1,
  nLoc = 1,
  nYear = 1,
  beta = NA,
  X = NA,
  locBoolean = NA
)
}
\arguments{
\item{phi}{Survival by location}

\item{p}{Detection probability by location}

\item{nInd}{Number of individuals}

\item{nLoc}{Number of locations}

\item{nYear}{Number of years of data}

\item{beta}{Covariate parameters}

\item{X}{Covariates}
}
\value{
A character string, capitalized to title case.
}
\description{
Produces a tagged list of CJS objects
}
\examples{
hello("james bond")
}
