% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_duplicates.R
\name{find_duplicates}
\alias{find_duplicates}
\title{Find duplicates}
\usage{
find_duplicates(df, ...)
}
\arguments{
\item{df}{Data frame}

\item{...}{Unquoted variable names to look for dupes}
}
\description{
Find duplicates
}
